fix(addon): decouple addon with provider
