+++ pkg/platform/util/addon/readme.md
@@ -1,0 +1,5 @@
+# Introduction
+
+This package just for independent addon controller and decouple with provider.
+
+Not for platform controller, which has used providers.
